+++ public/login.html
@@ -6,11 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>html til presentasjons app</title>
     <link rel="stylesheet" href="styles/PresentasjonStyle.css" type="text/css">
-<<<<<<< HEAD
-
-=======
     
->>>>>>> 0f3e85139b80a894f204a77a455c917434067817
 </head>
 
 <body>
