+++ server.js
@@ -27,8 +27,11 @@ app.post('/signUp', function (request, response) {
   // Sends object to pool.js-->DB;
   try {
     pool.newUser(request.body);
+    response.status(200).send('User created');
+    return;
   } catch (error) {
     response.send(error + ' Try again');
+    return;
   }
   //let newUser = pool.newUser(data);
   // console.log("new user: " + newUser);
